% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_class.R
\name{getDescription}
\alias{getDescription}
\alias{getDescription,gene-method}
\title{getDescription}
\usage{
getDescription(object)

\S4method{getDescription}{gene}(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The description of the gene
}
\description{
This is a method of the "gene" class, called "getDescription", and it allows
the user to extract the description of a specific gene.
}
\examples{
# Because the gene class is a VIRTUAL class, for the example, an object of
# class "proteinCodingGene" has been used
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(
             id = "ENSP01234567891",
             protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
             description = "It has a specific function"),
       protein(
             id = "ENSP19876543210", 
             protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."), 
             description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
getDescription(gene1)
}
