% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_class.R
\name{getGeneId}
\alias{getGeneId}
\alias{getGeneId,gene-method}
\title{getGeneId}
\usage{
getGeneId(object)

\S4method{getGeneId}{gene}(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The Ensembl ID of the gene
}
\description{
This is a method of the "gene" class that can be used to access information
about the Ensembl ID gene. The get method is the only method specified for
the Ensembl ID gene because usually the ID of a gene should not change, even
if the gene is updated.
}
\examples{
# Because the gene class is a VIRTUAL class, for the example, an object of
# class "proteinCodingGene" has been used.
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(id = "ENSP01234567891", protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."), description = "It has a specific function"),
       protein(id = "ENSP19876543210", protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."), description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
# getGeneId(gene1)
}
