% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microRNAGene_class.R
\name{getMicroRNA}
\alias{getMicroRNA}
\alias{getMicroRNA,MicroRNAGene-method}
\title{getMicroRNA}
\usage{
getMicroRNA(object)

\S4method{getMicroRNA}{MicroRNAGene}(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The silencing mechanisms of all the possible microRNA
}
\description{
This is a method of the class "MicroRNAGene" class. It allows the users to
extract and inspect the information about the microRNAs. This object will
contain the information about the ID, sequence and silencing mechanism
of each microRNA defined for a specific gene.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- microRNAGene(
   microRNAs = list(
       microRNA(id = "ENST01234567891",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has .."),
       microRNA(id = "ENST19876543210",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has ..")),
   geneID = "ENSG01234567891",
   gene_symbol = "MIR 12",
   full_gene_name = "microRNA 23",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
getMicroRNA(gene1)
}
