% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinCodingGene_class.R
\name{getProteins}
\alias{getProteins}
\alias{getProteins,ProteinCodingGene-method}
\title{getProteins}
\usage{
getProteins(object)

\S4method{getProteins}{ProteinCodingGene}(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The protein function of all the possible isoform of a gene
}
\description{
This is a method of the class "ProteinCodingGene" class. It allows
the user to acccess the information about the protein(s) codified
by the gene. This object will contain the information about the ID,
sequence and description of each protein defined for a specific gene.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(id = "ENSP01234567891",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function"),
       protein(id = "ENSP19876543210",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
getProteins(gene1)
}
