% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRNA_class.R
\name{getRRNA}
\alias{getRRNA}
\alias{getRRNA,rRNAgene-method}
\title{getRRNA}
\usage{
getRRNA(object)

\S4method{getRRNA}{rRNAgene}(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The information about the rRNA
}
\description{
This is a method of the class "rRNAgene" class. It allows the users to
extract and inspect the information about the rRNA. This object will
contain the information about the ID and ribosomal subunit of the rRNA
for a specific gene.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- rRNAgene(
   rRNA = rRNA(
         id = "ENST01234567891",
         rRNA_sequence = RNAString("AUUUGA"),
         ribosomal_subunit = "28S"),
   geneID = "ENSG01234567891",
   gene_symbol = "RNA28S",
   full_gene_name = "RNA, 28S ribosomal",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
getRRNA(gene1)
}
