% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_class.R
\name{getStructure}
\alias{getStructure}
\alias{getStructure,gene-method}
\title{getStructure}
\usage{
getStructure(object)

\S4method{getStructure}{gene}(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The information about the structure of the gene
}
\description{
This is a method of the "gene" class, called "getStructure", and it allows
the user to extract from a specific object all the information on the
structure of the gene.
}
\examples{
# Because the gene class is a VIRTUAL class, for the example, an object of
# class "proteinCodingGene" has been used
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(
             id = "ENSP01234567891", 
             protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."), 
             description = "It has a specific function"),
       protein(
             id = "ENSP19876543210", 
             protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."), 
             description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
getStructure(gene1)
}
