% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_class.R
\name{getTRNA}
\alias{getTRNA}
\alias{getTRNA,tRNAgene-method}
\title{getTRNA}
\usage{
getTRNA(object)

\S4method{getTRNA}{tRNAgene}(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The information about the tRNA
}
\description{
This is a method of the class "tRNAgene" class. It allows the user to acccess
the information about the tRNA codified by the gene. This object will contain
the information about the ID, sequence of the tRNA and the aminoacid and
anticodon carried by it.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- tRNAgene(
       tRNA = tRNA(
           id = "ENST01987654321",
           tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
           aminoacid = AAString("A"),
           anticodon = RNAString("AGC")),
       geneID = "ENSG01234567891",
       gene_symbol = "TRA-AGC1-1",
       full_gene_name = "tRNA-ALAAGC 1-1",
       description = "This gene has a specific function",
       structure = GRanges("chr1:1-1000"))
getTRNA(gene1)
}
