% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longNonCodingGene_class.R
\name{lengthProductLNCRG}
\alias{lengthProductLNCRG}
\title{lengthProductLNCRG}
\usage{
lengthProductLNCRG(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The length of the long non coding RNA sequence
}
\description{
This function helps in calculating the length of the RNA sequence
of all the long non coding RNA that are codified by the gene
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- longNonCodingRNAGene(
   LongNonCodingRNAs = list(
       longNonCodingRNA(
             id = "ENST01234567891",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism"),
       longNonCodingRNA(
             id = "ENST19876543210",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism")),
   geneID = "ENSG01234567891",
   gene_symbol = "LINC01018",
   full_gene_name = "long intergenic non-protein coding RNA 01018",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
lengthProductLNCRG(gene1)
}
