% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microRNAGene_class.R
\name{lengthProductMRG}
\alias{lengthProductMRG}
\title{lengthProductMRG}
\usage{
lengthProductMRG(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The length of the microRNAs
}
\description{
This function helps in calculating the length of the RNA sequence
of all the microRNA that are codified by the gene
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- microRNAGene(
   microRNAs = list(
       microRNA(id = "ENST01234567891",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has .."),
       microRNA(id = "ENST19876543210",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has ..")),
   geneID = "ENSG01234567891",
   gene_symbol = "MIR 12",
   full_gene_name = "microRNA 23",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
lengthProductMRG(gene1)
}
