% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinCodingGene_class.R
\name{lengthProductPCG}
\alias{lengthProductPCG}
\title{lengthProductPCG}
\usage{
lengthProductPCG(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The length of the aminoacid sequence of the product(s) of the gene
}
\description{
This function helps in calculating the length of the aminoacid
sequence of all the proteins that are codified by the gene.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(id = "ENSP01234567891",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function"),
       protein(id = "ENSP19876543210",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
lengthProductPCG(gene1)
}
