% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRNA_class.R
\name{lengthProductRRG}
\alias{lengthProductRRG}
\title{lengthProductRRG}
\usage{
lengthProductRRG(gene)
}
\arguments{
\item{gene}{A gene}
}
\value{
The length of the rRNA sequence
}
\description{
This function helps in calculating the length of the RNA sequence of the
rRNA that is codified by the gene.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- rRNAgene(
   rRNA = rRNA(
         id = "ENST01234567891",
         rRNA_sequence = RNAString("AUUUGA"),
         ribosomal_subunit = "28S"),
   geneID = "ENSG01234567891",
   gene_symbol = "RNA28S",
   full_gene_name = "RNA, 28S ribosomal",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
lengthProductRRG(gene1)
}
