% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_class.R
\name{lengthProductTRG}
\alias{lengthProductTRG}
\title{lengthProductTRG}
\usage{
lengthProductTRG(object)
}
\arguments{
\item{object}{A gene}
}
\value{
The length of the microRNA sequence as the product(s) of the gene
}
\description{
This function helps in calculating the length of the RNA sequence
of all the tRNA that are codified by the gene
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- tRNAgene(
       tRNA = tRNA(
           id = "ENST01987654321",
           tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
           aminoacid = AAString("A"),
           anticodon = RNAString("AGC")),
       geneID = "ENSG01234567891",
       gene_symbol = "TRA-AGC1-1",
       full_gene_name = "tRNA-ALAAGC 1-1",
       description = "This gene has a specific function",
       structure = GRanges("chr1:1-1000"))
lengthProductTRG(gene1)
}
