% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_class.R
\docType{class}
\name{LongNonCodingRNA-class}
\alias{LongNonCodingRNA-class}
\alias{longNonCodingRNA}
\title{longNonCodingRNA class}
\arguments{
\item{id}{Ensembl ID of the long non coding RNA}

\item{long_non_coding_RNA_sequence}{Sequence of the long non coding RNA}

\item{regulatory_mechanism}{Description of the regulatory mechanism of the long non coding RNA}
}
\description{
This class is called "LongNonCodingRNA" and it inherits from the virtual
class "product". It represents the product of the Long Non Coding RNA genes.
}
\examples{
library(Biostrings)
longNonCodingRNA(
     id = "ENST01234567891",
     long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
     regulatory_mechanism = "Description of the regulatory mechanism")
}
