% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longNonCodingGene_class.R
\docType{class}
\name{longNonCodingRNAGene-class}
\alias{longNonCodingRNAGene-class}
\alias{longNonCodingRNAGene}
\title{Long non coding RNA gene class}
\arguments{
\item{geneID}{Ensembl gene ID}

\item{gene_symbol}{HUGO gene symbol}

\item{full_gene_name}{The full name of the gene}

\item{description}{A brief description of the gene}

\item{structure}{chromosome, start, stop, ect.}

\item{LongNonCodingRNAs}{Information about the long non coding RNA}
}
\description{
This class represents the genes whose transcript(s) are classified as long
non coding RNA. This class inherits all the attributes of the virtual class
"gene" and it furthermore specifies the type of the gene by adding
information about the long non coding RNA(s). The "LongNonCodingRNAs"
attribute is a list of elements of the class "LongNonCodingRNA". This
class inherits from the virtual class "product" and it has different
parameters such as the "id" of the long non coding RNA, the sequence of the
long non coding RNA ("long_non_coding_RNA_sequence") and lastly its
regulatory mechanism ("regulatory_mechanism").
}
\examples{
library(Biostrings)
library(GenomicRanges)
longNonCodingRNAGene(
   LongNonCodingRNAs = list(
       longNonCodingRNA(
             id = "ENST01234567891",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism"),
       longNonCodingRNA(
             id = "ENST19876543210",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism")),
   geneID = "ENSG01234567891",
   gene_symbol = "LINC01018",
   full_gene_name = "long intergenic non-protein coding RNA 01018",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
}
