% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_class.R
\docType{class}
\name{microRNA-class}
\alias{microRNA-class}
\alias{microRNA}
\title{microRNA class}
\arguments{
\item{id}{Ensembl ID of the MicroRNA}

\item{microRNA_sequence}{Sequence of the MicroRNA}

\item{silencing_mechanism}{Description of the silencing mechanism of the MicroRNA}
}
\description{
This class is called "MicroRNA" and it inherits from the virtual class
"product". It represents the product of the microRNA genes.
}
\examples{
library(Biostrings)
microRNA(
     id = "ENST01234567891",
     microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
     silencing_mechanism = "The microRNA has ..")
}
