% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microRNAGene_class.R
\docType{class}
\name{MicroRNAGene-class}
\alias{MicroRNAGene-class}
\alias{microRNAGene}
\title{microRNA gene class}
\arguments{
\item{geneID}{Ensembl gene ID}

\item{gene_symbol}{HUGO gene symbol}

\item{full_gene_name}{The full name of the gene}

\item{description}{A brief description of the gene}

\item{structure}{chromosome, start, stop, ect.}

\item{microRNAs}{microRNAs of a specific gene}
}
\description{
This class represents the genes whose transcript(s) are classified as
microRNA. This class inherits all the attributes of the virtual class
"gene" and it furthermore specifies the type of the gene by adding
information about the microRNA(s). The "microRNAs" attribute is a list
of elements of the class "microRNA". This class inherits from the virtual
class "product" and it has different parameters such as the "id" of the
microRNA, the sequence of the microRNA ("microRNA_sequence") and lastly
its silencing mechanism ("silencing_mechanism").
}
\examples{
library(Biostrings)
library(GenomicRanges)
microRNAGene(
   microRNAs = list(
       microRNA(id = "ENST01234567891",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has .."),
       microRNA(id = "ENST19876543210",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has ..")),
   geneID = "ENSG01234567891",
   gene_symbol = "MIR 12",
   full_gene_name = "microRNA 23",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
}
