% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_class.R
\docType{class}
\name{Product-class}
\alias{Product-class}
\title{product class}
\arguments{
\item{id}{Ensembl ID of the product of a gene}
}
\description{
This class is a virtual class for all the products of the genes specified
in each class. In this virtual class the only attribute specified is the
EnsemblID of the product of the genes. This is because it is the only
attribute common to all the products (proteis, microRNAs, lncRNAs, rRNAs,
tRNAs).
}
