% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_class.R
\docType{class}
\name{Protein-class}
\alias{Protein-class}
\alias{protein}
\title{protein Product Class}
\arguments{
\item{id}{Ensembl ID of the protein}

\item{protein_sequence}{Sequence of the protein}

\item{protein_function}{Description of the function of the protein}
}
\description{
This class is called "Protein" and it inherits from the virtual class
"product". It represents the product of the protein-coding genes.
}
\examples{
library(Biostrings)
protein(
     id = "ENSP01234567891",
     protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
     description = "It has a specific function")
}
