% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinCodingGene_class.R
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\alias{proteinCodingGene}
\title{protein coding gene class}
\arguments{
\item{geneID}{Ensembl gene ID}

\item{gene_symbol}{HUGO gene symbol}

\item{full_gene_name}{The full name of the gene}

\item{description}{A brief description of the gene}

\item{structure}{chromosome, start, stop, etc.}

\item{proteins}{list of all the protein codified by the gene}
}
\description{
This class represents the genes whose transcript(s) codify for proteins.
This class inherits all the attributes of the virtual class "gene" and
it furthermore specifies the type of the gene by adding information about
the protein(s), codified by this type of gene. The "proteins" attribute is
a list of elements of the class "Protein". This class inherits from the
virtual class "product" and it has different parameters such as the "id"
of the protein, the sequence of the protein ("protein_sequence") and lastly
the description of the protein ("description").
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(id = "ENSP01234567891",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function"),
       protein(id = "ENSP19876543210",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
}
