% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_class.R
\docType{class}
\name{rRNA-class}
\alias{rRNA-class}
\alias{rRNA}
\title{rRNA class}
\arguments{
\item{id}{Ensembl ID of the rRNA}

\item{microRNA_sequence}{Sequence of the rRNA}

\item{ribosomial_subunit}{The ribosomal subunit}
}
\description{
This class is called "rRNA" and it inherits from the virtual
class "product". It represents the product of the rRNA genes.
}
\examples{
library(Biostrings)
rRNA(
     id = "ENST01234567891",
     rRNA_sequence = RNAString("AUUUGA"),
     ribosomal_subunit = "28S")
}
