% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRNA_class.R
\docType{class}
\name{rRNAgene-class}
\alias{rRNAgene-class}
\alias{rRNAgene}
\title{rRNA gene class}
\arguments{
\item{geneID}{Ensembl gene ID}

\item{gene_symbol}{HUGO gene symbol}

\item{full_gene_name}{The full name of the gene}

\item{description}{A brief description of the gene}

\item{structure}{chromosome, start, stop, etc.}

\item{rRNA}{All the information about the rRNA}
}
\description{
This class rapresents the genes whose transcript(s) are classified as rRNA
(ribosomal RNA). This class inherits all the attributes of the virtual class
"gene" and it furthermore specifies the type of the gene by adding
information about the rRNA. The "rRNA" attribute is a single object of class
"rRNA". This class inherits from the virtual class "product" and it has
different parameters such as the "id" of the rRNA, the sequence of the rRNA
("rRNA_sequence") and lastly its ribosomal subunit ("ribosomal_subunit").
}
\examples{
library(Biostrings)
library(GenomicRanges)
rRNAgene(
   rRNA = rRNA(
         id = "ENST01234567891",
         rRNA_sequence = RNAString("AUUUGA"),
         ribosomal_subunit = "28S"),
   geneID = "ENSG01234567891",
   gene_symbol = "RNA28S",
   full_gene_name = "RNA, 28S ribosomal",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
}
