% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_class.R
\name{setAminoacid}
\alias{setAminoacid}
\alias{setAminoacid,tRNAgene-method}
\title{setAminoacid}
\usage{
setAminoacid(object, new_aminoacid)

\S4method{setAminoacid}{tRNAgene}(object, new_aminoacid)
}
\arguments{
\item{object}{A tRNA gene gene}

\item{new_aminoacid}{A aminoacid for the tRNA gene}
}
\value{
the gene with the tRNA with a new aminoacid
}
\description{
This is a method of the class "tRNAgene". The new aminoacid has to respect
different conditions, for example it has to be of class AAString and also
has to be a valid aminoacid. If this is not the case, then the following
will raise: "Invalid aminoacid".
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- tRNAgene(
       tRNA = tRNA(
           id = "ENST01987654321",
           tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
           aminoacid = AAString("A"), anticodon = RNAString("AGC")),
       geneID = "ENSG01234567891",
       gene_symbol = "TRA-AGC1-1",
       full_gene_name = "tRNA-ALAAGC 1-1",
       description = "This gene has a specific function",
       structure = GRanges("chr1:1-1000"))
gene1 <- setAminoacid(gene1, AAString("M"))
}
