% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_class.R
\name{setAnticodon}
\alias{setAnticodon}
\alias{setAnticodon,tRNAgene-method}
\title{setAnticodon}
\usage{
setAnticodon(object, new_anticodon)

\S4method{setAnticodon}{tRNAgene}(object, new_anticodon)
}
\arguments{
\item{object}{A tRNA gene gene}

\item{new_anticodon}{An anticodon for the tRNA gene}
}
\value{
the gene with the tRNA with a new anticodon
}
\description{
This is a method of the class "tRNAgene". The new anticodon has to respect
different conditions, for example it has to be of class RNAString and also
its length has to be of 3 base. If this is not the case, then the following
will raise: "Invalid aminoacid".
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- tRNAgene(
       tRNA = tRNA(
           id = "ENST01987654321",
           tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
           aminoacid = AAString("A"),
           anticodon = RNAString("AGC")),
       geneID = "ENSG01234567891",
       gene_symbol = "TRA-AGC1-1",
       full_gene_name = "tRNA-ALAAGC 1-1",
       description = "This gene has a specific function",
       structure = GRanges("chr1:1-1000"))
gene1 <- setAnticodon(gene1, RNAString("AUU"))
}
