% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_class.R
\name{setDescription}
\alias{setDescription}
\alias{setDescription,gene-method}
\title{setDescription}
\usage{
setDescription(object, new_description)

\S4method{setDescription}{gene}(object, new_description)
}
\arguments{
\item{object}{A gene}

\item{new_description}{A new description of the gene}
}
\value{
The gene with the new description
}
\description{
This is a method of the "gene" class, called "setDescription", and it allows
the user to change the description of a specific gene.
}
\examples{
# Because the gene class is a VIRTUAL class, for the example, an object of
# class "proteinCodingGene" has been used
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(
             id = "ENSP01234567891",
             protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
             description = "It has a specific function"),
       protein(
             id = "ENSP19876543210",
             protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
             description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
setDescription(gene1, "new description")
}
