% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_class.R, R/longNonCodingGene_class.R,
%   R/microRNAGene_class.R, R/proteinCodingGene_class.R, R/rRNA_class.R,
%   R/tRNA_class.R
\name{setFullName}
\alias{setFullName}
\alias{setFullName,gene-method}
\alias{setFullName,longNonCodingRNAGene-method}
\alias{setFullName,MicroRNAGene-method}
\alias{setFullName,ProteinCodingGene-method}
\alias{setFullName,rRNAgene-method}
\alias{setFullName,tRNAgene-method}
\title{setFullName}
\usage{
setFullName(object, new_gene_full_name)

\S4method{setFullName}{gene}(object, new_gene_full_name)

\S4method{setFullName}{longNonCodingRNAGene}(object, new_gene_full_name)

\S4method{setFullName}{MicroRNAGene}(object, new_gene_full_name)

\S4method{setFullName}{ProteinCodingGene}(object, new_gene_full_name)

\S4method{setFullName}{rRNAgene}(object, new_gene_full_name)

\S4method{setFullName}{tRNAgene}(object, new_gene_full_name)
}
\arguments{
\item{object}{An tRNA gene}

\item{new_gene_full_name}{The new full name of the gene}
}
\value{
The gene with the new full name

The gene with the new full gene name

The microRNA gene with the new full gene name

The gene with the new full name

The gene with the new full name

The gene with the new full name
}
\description{
This is a method of the "gene" class, called "setFullName", that allows the
user to change the information about the full name of the gene.

This method of class "longNonCodingRNAGene" which allows
the users to change the information about the
full name of the gene.

This is a method inherited from the "gene" class. For the class
"MicroRNAGene" the new full name has to respect different conditions
such as it has to be a character and also start with "microRNA"
followed by a space and then a number.

This is a method inherited from the "gene" class. For the class
"ProteinCodingGene" there have been added additional constrains
regarding the new full gene name. Indeed it must be of class
"character" and also must be uppercase.

This is a method inherited from the "gene" class. For the class
"rRNAgene" there have been added additional constraints
regarding the new full gene name. Indeed it must be of class
"character" and also must have a specific structure.

This is a method inherited from the "gene" class. For the class
"tRNAgene" there have been added additional constrains
regarding the new full gene name. Indeed it must be of class
"character" and also must have a specific structure.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- longNonCodingRNAGene(
   LongNonCodingRNAs = list(
       longNonCodingRNA(
             id = "ENST01234567891",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism"),
       longNonCodingRNA(
             id = "ENST19876543210",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism")),
   geneID = "ENSG01234567891",
   gene_symbol = "LINC01018",
   full_gene_name = "long intergenic non-protein coding RNA 01018",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
gene1 <- setFullName(gene1, "long intergenic non-protein coding RNA 01234")
library(Biostrings)
library(GenomicRanges)
gene1 <- microRNAGene(
   microRNAs = list(
       microRNA(id = "ENST01234567891",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has .."),
       microRNA(id = "ENST19876543210",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has ..")),
   geneID = "ENSG01234567891",
   gene_symbol = "MIR 12",
   full_gene_name = "microRNA 23",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
gene1 <- setFullName(gene1, "microRNA 123")
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(id = "ENSP01234567891",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function"),
       protein(id = "ENSP19876543210",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
gene1 <- setFullName(gene1, "GENE TWO")
library(Biostrings)
library(GenomicRanges)
gene1 <- rRNAgene(
   rRNA = rRNA(
         id = "ENST01234567891",
         rRNA_sequence = RNAString("AUUUGA"),
         ribosomal_subunit = "28S"),
   geneID = "ENSG01234567891",
   gene_symbol = "RNA28S",
   full_gene_name = "RNA, 28S ribosomal",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
gene1 <- setFullName(gene1, "RNA, 45S ribosomal")
library(Biostrings)
library(GenomicRanges)
gene1 <- tRNAgene(
       tRNA = tRNA(
           id = "ENST01987654321",
           tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
           aminoacid = AAString("A"), anticodon = RNAString("AGC")),
       geneID = "ENSG01234567891",
       gene_symbol = "TRA-AGC1-1",
       full_gene_name = "tRNA-ALAAGC 1-1",
       description = "This gene has a specific function",
       structure = GRanges("chr1:1-1000"))
gene1 <- setFullName(gene1, "tRNA-METAUU 1-1")
}
