% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longNonCodingGene_class.R
\name{setlongNonCodingRNA}
\alias{setlongNonCodingRNA}
\alias{setlongNonCodingRNA,longNonCodingRNAGene-method}
\title{setlongNonCodingRNA}
\usage{
setlongNonCodingRNA(object, lncRNA_EnsemblID, new_lncRNA)

\S4method{setlongNonCodingRNA}{longNonCodingRNAGene}(object, lncRNA_EnsemblID, new_lncRNA)
}
\arguments{
\item{object}{A gene}

\item{lncRNA_EnsemblID}{The EnsemblID of the long non coding RNA that the user wants to change}

\item{new_lncRNA}{The new long non coidng RNA}
}
\value{
The gene with the new long non coding RNA
}
\description{
This method of class "longNonCodingRNAGene" which allows the users
to change the information about the long Non Coding RNAs.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- longNonCodingRNAGene(
   LongNonCodingRNAs = list(
       longNonCodingRNA(
             id = "ENST01234567891",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism"),
       longNonCodingRNA(
             id = "ENST19876543210",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism")),
   geneID = "ENSG01234567891",
   gene_symbol = "LINC01018",
   full_gene_name = "long intergenic non-protein coding RNA 01018",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
setlongNonCodingRNA(
         gene1, "ENST01234567891",
         longNonCodingRNA(
             id = "ENST00000000001",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism"))
}
