% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microRNAGene_class.R
\name{setMicroRNA}
\alias{setMicroRNA}
\alias{setMicroRNA,MicroRNAGene-method}
\title{setMicroRNA}
\usage{
setMicroRNA(object, miRNA_EnsemblID, new_microRNA)

\S4method{setMicroRNA}{MicroRNAGene}(object, miRNA_EnsemblID, new_microRNA)
}
\arguments{
\item{object}{A gene}

\item{miRNA_EnsemblID}{The EnsemblID of the microRNA for which the user wants to chaange the information}

\item{new_microRNA}{The new microRNA}
}
\value{
The gene with the new microRNA
}
\description{
This method is called "setMicroRNA" which allows the users to change the
information about a specific microRNA. The user can change the information
about the ID, sequence and silencing mechanism of the microRNA.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- microRNAGene(
   microRNAs = list(
       microRNA(id = "ENST01234567891",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has .."),
       microRNA(id = "ENST19876543210",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has ..")),
   geneID = "ENSG01234567891",
   gene_symbol = "MIR 12",
   full_gene_name = "microRNA 23",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
gene1 <- setMicroRNA(
     gene1, "ENST01234567891",
     microRNA(
         id = "ENST00000000001",
         microRNA_sequence = RNAString("ACG"),
         silencing_mechanism = "Its silencing mechanism consists of")
)
}
