% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinCodingGene_class.R
\name{setProtein}
\alias{setProtein}
\alias{setProtein,ProteinCodingGene-method}
\title{setProtein}
\usage{
setProtein(object, protein_EnsemblID, new_protein)

\S4method{setProtein}{ProteinCodingGene}(object, protein_EnsemblID, new_protein)
}
\arguments{
\item{object}{A gene}

\item{protein_EnsemblID}{EnsemblID of the protein for which the user wants to change the function}

\item{new_protein}{The new function for the protein}
}
\value{
The protein gene with a new protein
}
\description{
This is a method of the class "ProteinCodingGene" class. It allows
the user to modify the information about a specific protein codified
by the gene.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(id = "ENSP01234567891",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function"),
       protein(id = "ENSP19876543210",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
gene1 <- setProtein(gene1, "ENSP01234567891",
     protein(id = "ENSP00000000001",
     protein_sequence = AAString("ARNDCQ"),
     description = "It has no specific function"))
}
