% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRNA_class.R
\name{setRibosomalSub}
\alias{setRibosomalSub}
\alias{setRibosomalSub,rRNAgene-method}
\title{setRibosomalSub}
\usage{
setRibosomalSub(object, new_RibosomalSub)

\S4method{setRibosomalSub}{rRNAgene}(object, new_RibosomalSub)
}
\arguments{
\item{object}{A rRNA gene gene}

\item{new_RibosomalSub}{A new ribosomal subunit for the rRNA gene}
}
\value{
the gene with the new informations
}
\description{
This is a method inherited from the "gene" class. For the class "rRNAgene"
the new ribosomal subunit has to respect different conditions, for example
it has to be a character object and also be a valid ribosomal subunits.
If this is not the case, then the following will raise: "Invalid Ribosomal
Subunit".
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- rRNAgene(
   rRNA = rRNA(
         id = "ENST01234567891",
         rRNA_sequence = RNAString("AUUUGA"),
         ribosomal_subunit = "28S"),
   geneID = "ENSG01234567891",
   gene_symbol = "RNA28S",
   full_gene_name = "RNA, 28S ribosomal",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
gene1 <- setRibosomalSub(gene1, "45S")
}
