% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRNA_class.R
\name{setRRNA}
\alias{setRRNA}
\alias{setRRNA,rRNAgene-method}
\title{setRRNA}
\usage{
setRRNA(object, new_rRNA)

\S4method{setRRNA}{rRNAgene}(object, new_rRNA)
}
\arguments{
\item{object}{A gene}

\item{new_rRNA}{The new rRNA}
}
\value{
A gene with a new rRNA
}
\description{
This method is called "setRRNA" which allows the users to change the
information about a specific rRNA. The user can change the information
about the ID, sequence and silencing mechanism of the rRNA.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- rRNAgene(
   rRNA = rRNA(
         id = "ENST01234567891",
         rRNA_sequence = RNAString("AUUUGA"),
         ribosomal_subunit = "28S"),
   geneID = "ENSG01234567891",
   gene_symbol = "RNA28S",
   full_gene_name = "RNA, 28S ribosomal",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
setRRNA(gene1, rRNA(id = "ENST00000000001",
     rRNA_sequence = RNAString("AUUUGA"),
     ribosomal_subunit = "45S"))
}
