% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_class.R
\name{setStructure}
\alias{setStructure}
\alias{setStructure,gene-method}
\title{setStructure}
\usage{
setStructure(object, new_structure)

\S4method{setStructure}{gene}(object, new_structure)
}
\arguments{
\item{object}{A gene}

\item{new_structure}{The new structure of the gene}
}
\value{
The gene with the new structure
}
\description{
This is a method of the "gene" class, called "setStructure", that is useful
to extract from a specific gene all the information about the structure.
}
\examples{
# Because the gene class is a VIRTUAL class, for the example, an object of
# class "proteinCodingGene" has been used
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(
             id = "ENSP01234567891",
             protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
             description = "It has a specific function"),
       protein(
             id = "ENSP19876543210",
             protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
             description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
setStructure(gene1, GRanges("chr1:1-100"))
}
