% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_class.R, R/longNonCodingGene_class.R,
%   R/microRNAGene_class.R, R/proteinCodingGene_class.R, R/rRNA_class.R,
%   R/tRNA_class.R
\name{setSymbol}
\alias{setSymbol}
\alias{setSymbol,gene-method}
\alias{setSymbol,longNonCodingRNAGene-method}
\alias{setSymbol,MicroRNAGene-method}
\alias{setSymbol,ProteinCodingGene-method}
\alias{setSymbol,rRNAgene-method}
\alias{setSymbol,tRNAgene-method}
\title{setSymbol}
\usage{
setSymbol(object, new_gene_symbol)

\S4method{setSymbol}{gene}(object, new_gene_symbol)

\S4method{setSymbol}{longNonCodingRNAGene}(object, new_gene_symbol)

\S4method{setSymbol}{MicroRNAGene}(object, new_gene_symbol)

\S4method{setSymbol}{ProteinCodingGene}(object, new_gene_symbol)

\S4method{setSymbol}{rRNAgene}(object, new_gene_symbol)

\S4method{setSymbol}{tRNAgene}(object, new_gene_symbol)
}
\arguments{
\item{object}{An tRNA gene}

\item{new_gene_symbol}{A new symbol for a specific gene}
}
\value{
The gene with the new HUGO gene symbol

The gene with the new HUGO gene symbol

The microRNA gene with the new HUGO gene symbol

The gene with the new HUGO gene symbol

The gene with the new HUGO gene symbol

The gene with the new HUGO gene symbol
}
\description{
This is a method of the "gene" class called "setSymbol" that can be used to
change the information about the HUGO gene symbol.

This method of class "longNonCodingRNAGene" which allows
the users to change the information about the
HUGO gene symbol.

This is a method inherited from the "gene" class. For the class
"MicroRNAGene" the new gene symbol has to respect different
conditions, for example it has to be a character object and also
start with "MIR " followed by a number. If this is not the case,
then the following will raise: "Invalid gene symbol".

This is a method inherited from the "gene" class. For the class
"ProteinCodingGene" there have been added additional constrains
regarding the new HUGO gene symbol. Indeed it must be of class
"character" and also must be uppercase.

This is a method inherited from the "gene" class. For the class
"rRNAgene" there have been added additional constraints
regarding the new HUGO gene symbol. Indeed it must be of class
"character" and also must have a specific structure.

This is a method inherited from the "gene" class. For the class
"tRNAgene" there have been added additional constrains
regarding the new HUGO gene symbol. Indeed it must be of class
"character" and also must have a specif structure.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- longNonCodingRNAGene(
   LongNonCodingRNAs = list(
       longNonCodingRNA(
             id = "ENST01234567891",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism"),
       longNonCodingRNA(
             id = "ENST19876543210",
             long_non_coding_RNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
             regulatory_mechanism = "Description of the specific regulatory mechanism")),
   geneID = "ENSG01234567891",
   gene_symbol = "LINC01018",
   full_gene_name = "long intergenic non-protein coding RNA 01018",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
gene1 <- setSymbol(gene1, "LINC01010")
library(Biostrings)
library(GenomicRanges)
gene1 <- microRNAGene(
   microRNAs = list(
       microRNA(id = "ENST01234567891",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has .."),
       microRNA(id = "ENST19876543210",
                 microRNA_sequence = RNAString("ACGUMRWSYKVHDBN-+."),
                 silencing_mechanism = "The microRNA has ..")),
   geneID = "ENSG01234567891",
   gene_symbol = "MIR 12",
   full_gene_name = "microRNA 23",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
gene1 <- setSymbol(gene1, "MIR 123")
library(Biostrings)
library(GenomicRanges)
gene1 <- proteinCodingGene(
   proteins = list(
       protein(id = "ENSP01234567891",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function"),
       protein(id = "ENSP19876543210",
                 protein_sequence = AAString("ARNDCQEGHILKMFPSTWYVUOBJZX*-+."),
                 description = "It has a specific function")),
   geneID = "ENSG01234567891",
   gene_symbol = "GENE1",
   full_gene_name = "GENE ONE",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000")
)
gene1 <- setSymbol(gene1, "GENE1")
library(Biostrings)
library(GenomicRanges)
gene1 <- rRNAgene(
   rRNA = rRNA(
         id = "ENST01234567891",
         rRNA_sequence = RNAString("AUUUGA"),
         ribosomal_subunit = "28S"),
   geneID = "ENSG01234567891",
   gene_symbol = "RNA28S",
   full_gene_name = "RNA, 28S ribosomal",
   description = "This gene has a specific function",
   structure = GRanges("chr1:1-1000"))
setRibosomalSub(gene1, "45S")
gene1 <- setSymbol(gene1, "RNA45S")
library(Biostrings)
library(GenomicRanges)
gene1 <- tRNAgene(
       tRNA = tRNA(
           id = "ENST01987654321",
           tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
           aminoacid = AAString("A"), anticodon = RNAString("AGC")),
       geneID = "ENSG01234567891",
       gene_symbol = "TRA-AGC1-1",
       full_gene_name = "tRNA-ALAAGC 1-1",
       description = "This gene has a specific function",
       structure = GRanges("chr1:1-1000"))
gene1 <- setSymbol(gene1, "TRM-AUU1-1")
}
