% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_class.R
\name{setTRNA}
\alias{setTRNA}
\alias{setTRNA,tRNAgene-method}
\title{setTRNA}
\usage{
setTRNA(object, new_tRNA)

\S4method{setTRNA}{tRNAgene}(object, new_tRNA)
}
\arguments{
\item{object}{A gene}

\item{new_tRNA}{The new tRNA}
}
\value{
The gene with the new tRNA
}
\description{
This is a method of the class "tRNAgene" class. It allows the user to change
the information about the tRNA. The user can change the information about the
ID, sequence, aminoacid and anticodon of the tRNA.
}
\examples{
library(Biostrings)
library(GenomicRanges)
gene1 <- tRNAgene(
       tRNA = tRNA(
           id = "ENST01987654321",
           tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
           aminoacid = AAString("A"),
           anticodon = RNAString("AGC")),
       geneID = "ENSG01234567891",
       gene_symbol = "TRA-AGC1-1",
       full_gene_name = "tRNA-ALAAGC 1-1",
       description = "This gene has a specific function",
       structure = GRanges("chr1:1-1000"))
gene1 <- setTRNA(gene1, tRNA(id = "ENST00000000001",
     tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
     aminoacid = AAString("M"),
     anticodon = RNAString("AGC")))
}
