% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_class.R
\docType{class}
\name{tRNA-class}
\alias{tRNA-class}
\alias{tRNA}
\title{tRNA class}
\arguments{
\item{id}{Ensembl ID of the tRNA}

\item{tRNA_sequence}{Sequence of the tRNA}

\item{aminoacid}{The aminoacid carried by the tRNA}

\item{anticodon}{The anticodon on the tRNA}
}
\description{
This class is called "tRNA" and it inherits from the virtual class
"product". It represents the product of the tRNA genes.
}
\examples{
library(Biostrings)
tRNA(
     id = "ENST00000000001",
     tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
     aminoacid = AAString("M"),
     anticodon = RNAString("AGC"))
}
