% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_class.R
\docType{class}
\name{tRNAgene-class}
\alias{tRNAgene-class}
\alias{tRNAgene}
\title{tRNA gene class}
\arguments{
\item{geneID}{Ensembl gene ID}

\item{gene_symbol}{HUGO gene symbol}

\item{full_gene_name}{The full name of the gene}

\item{description}{A brief description of the gene}

\item{structure}{chromosome, start, stop, etc.}

\item{tRNA}{All the information about the tRNA}
}
\description{
This class rapresents the genes whose transcript(s) are classified as tRNA
(transfer RNA). This class inherits all the attributes of the virtual class
"gene" and it furthermore specifies the type of the gene by adding
information about the tRNA. The "tRNA" attribute is a single object of class
"tRNA". This class inherits from the virtual class "product" and it has
different parameters such as the "id" of the tRNA, the sequence of the tRNA
("tRNA_sequence"), the aminoacid carried by the tRNA ("aminoacid") and the
anticodon sequence ("anticodon").
}
\examples{
library(Biostrings)
library(GenomicRanges)
tRNAgene(
       tRNA = tRNA(
           id = "ENST01987654321",
           tRNA_sequence = RNAString("AGCCGUMRWSYKVHDBN-+."),
           aminoacid = AAString("A"),
           anticodon = RNAString("AGC")),
       geneID = "ENSG01234567891",
       gene_symbol = "TRA-AGC1-1",
       full_gene_name = "tRNA-ALAAGC 1-1",
       description = "This gene has a specific function",
       structure = GRanges("chr1:1-1000"))
}
